+++ src/_refs.d.ts
@@ -1,6 +1,0 @@
-﻿///<reference path="./typings/asyncawait.d.ts" />
-///<reference path="./typings/bluebird.d.ts" />
-///<reference path="./typings/lodash.d.ts" />
-///<reference path="./typings/node.d.ts" />
-///<reference path="./typings/node-fibers.d.ts" />
-declare module "_refs" { }