+++ src/references.d.ts
@@ -1,0 +1,7 @@
+﻿/// <reference path="typings/asyncawait/asyncawait.d.ts" />
+/// <reference path="typings/bluebird/bluebird.d.ts" />
+/// <reference path="typings/chai/chai.d.ts" />
+/// <reference path="typings/lodash/lodash.d.ts" />
+/// <reference path="typings/mocha/mocha.d.ts" />
+/// <reference path="typings/node/node.d.ts" />
+/// <reference path="typings/node-fibers/node-fibers.d.ts" />